add argument equalCharge
Add argument equalCharge to compare only the ion mobility of equally charged
ions. The correct way would be to go back to the pep3d file and fetch the ion
mobility for the corresponding charge state from there. But because the current
results are not satisfying it seems to be not worth the additional effort.